requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Jinja2==3.1.4
python-pptx==0.6.23
requests==2.32.3